events created with foreign user key
