+++ client/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "axios": "^0.19.0",
     "firebase": "^7.2.2",
+    "mongojs": "^3.1.0",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-redux": "^7.1.3",
