+++ controllers/eventsController.js
@@ -5,7 +5,7 @@ module.exports = {
     findAll: function (req, res) {
         db.Event
             .find(req.query)
-            .sort({ date: -1 })
+            .populate("user")
             .then(dbModel => res.json(dbModel))
             .catch(err => res.status(422).json(err));
     },
