+++ models/events.js
@@ -2,6 +2,10 @@ const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 const eventSchema = new Schema({
+    user: {
+        type: Schema.Types.ObjectId,
+        ref: "User"
+    },
     userName: { type: String, required: true },
     roomName: { type: String, required: true },
     startTime: { type: String, required: true },
