+++ routes/api/index.js
@@ -3,6 +3,7 @@ const userRoutes = require("./users");
 const locationRoutes = require("./locations");
 const featureRoutes = require("./features");
 const roomRoutes = require("./rooms");
+const eventRoutes = require("./events");
 
 // User routes
 router.use("/users", userRoutes);
@@ -15,5 +16,8 @@ router.use("/features", featureRoutes)
 
 // Room routes
 router.use("/rooms", roomRoutes)
+
+// Event routes
+router.use("/events", eventRoutes)
 
 module.exports = router;
